LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/credentialless/dedicated-worker.tentative.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: ReferenceError: Can't find variable: worker
CONSOLE MESSAGE: Unhandled Promise Rejection: ReferenceError: Can't find variable: worker
CONSOLE MESSAGE: Unhandled Promise Rejection: ReferenceError: Can't find variable: worker
CONSOLE MESSAGE: Unhandled Promise Rejection: ReferenceError: Can't find variable: worker

PASS dedicated-worker
PASS fetch same-origin + credentialless worker
FAIL fetch same-origin assert_equals: coep:credentialless =>  expected "Worker blocked" but got "same_origin"
FAIL fetch cross-origin assert_equals: coep:credentialless =>  expected "Worker blocked" but got "cross_origin"
FAIL fetch cross-origin + credentialless worker assert_equals: coep:none =>  expected (undefined) undefined but got (string) "cross_origin"
